backend/requirements.txt
django
psycopg2-binary
pandas
djangorestframework
django-extensions
django-cors-headers
djangorestframework-simplejwt
channels
daphne
paho-mqtt
openpyxl
reportlab
redis
django-redis
channels_redis
numpy
pandas
matplotlib
pytest
pytest-django
requests